+++ sdk/mpr/writer_microflow.go
@@ -155,7 +155,7 @@ func serializeSequenceFlow(flow *microflows.SequenceFlow) bson.D {
 		switch cv := flow.CaseValue.(type) {
 		case microflows.EnumerationCase:
 			caseValues = bson.A{
-				int32(2),
+				int64(2),
 				bson.D{
 					{Key: "$ID", Value: idToBsonBinary(string(cv.ID))},
 					{Key: "$Type", Value: "Microflows$EnumerationCase"},
@@ -164,7 +164,7 @@ func serializeSequenceFlow(flow *microflows.SequenceFlow) bson.D {
 			}
 		case *microflows.EnumerationCase:
 			caseValues = bson.A{
-				int32(2),
+				int64(2),
 				bson.D{
 					{Key: "$ID", Value: idToBsonBinary(string(cv.ID))},
 					{Key: "$Type", Value: "Microflows$EnumerationCase"},
@@ -178,7 +178,7 @@ func serializeSequenceFlow(flow *microflows.SequenceFlow) bson.D {
 		{Key: "$ID", Value: idToBsonBinary(string(flow.ID))},
 		{Key: "$Type", Value: "Microflows$SequenceFlow"},
 		{Key: "CaseValues", Value: caseValues},
-		{Key: "DestinationConnectionIndex", Value: int32(flow.DestinationConnectionIndex)},
+		{Key: "DestinationConnectionIndex", Value: int64(flow.DestinationConnectionIndex)},
 		{Key: "DestinationPointer", Value: idToBsonBinary(string(flow.DestinationID))},
 		{Key: "IsErrorHandler", Value: flow.IsErrorHandler},
 		{Key: "Line", Value: bson.D{
@@ -187,7 +187,7 @@ func serializeSequenceFlow(flow *microflows.SequenceFlow) bson.D {
 			{Key: "DestinationControlVector", Value: "0;0"},
 			{Key: "OriginControlVector", Value: "0;0"},
 		}},
-		{Key: "OriginConnectionIndex", Value: int32(flow.OriginConnectionIndex)},
+		{Key: "OriginConnectionIndex", Value: int64(flow.OriginConnectionIndex)},
 		{Key: "OriginPointer", Value: idToBsonBinary(string(flow.OriginID))},
 	}
 }
@@ -197,7 +197,7 @@ func serializeAnnotationFlow(af *microflows.AnnotationFlow) bson.D {
 	return bson.D{
 		{Key: "$ID", Value: idToBsonBinary(string(af.ID))},
 		{Key: "$Type", Value: "Microflows$AnnotationFlow"},
-		{Key: "DestinationConnectionIndex", Value: int32(0)},
+		{Key: "DestinationConnectionIndex", Value: int64(0)},
 		{Key: "DestinationPointer", Value: idToBsonBinary(string(af.DestinationID))},
 		{Key: "Line", Value: bson.D{
 			{Key: "$ID", Value: idToBsonBinary(generateUUID())},
@@ -205,7 +205,7 @@ func serializeAnnotationFlow(af *microflows.AnnotationFlow) bson.D {
 			{Key: "DestinationControlVector", Value: "0;0"},
 			{Key: "OriginControlVector", Value: "0;0"},
 		}},
-		{Key: "OriginConnectionIndex", Value: int32(0)},
+		{Key: "OriginConnectionIndex", Value: int64(0)},
 		{Key: "OriginPointer", Value: idToBsonBinary(string(af.OriginID))},
 	}
 }
@@ -527,8 +527,8 @@ func serializeMicroflowObject(obj microflows.MicroflowObject) bson.D {
 func serializePoint(pt model.Point) bson.D {
 	return bson.D{
 		{Key: "$Type", Value: "Common$Point"},
-		{Key: "X", Value: int32(pt.X)},
-		{Key: "Y", Value: int32(pt.Y)},
+		{Key: "X", Value: int64(pt.X)},
+		{Key: "Y", Value: int64(pt.Y)},
 	}
 }
 
@@ -536,8 +536,8 @@ func serializePoint(pt model.Point) bson.D {
 func serializeSize(sz model.Size) bson.D {
 	return bson.D{
 		{Key: "$Type", Value: "Common$Size"},
-		{Key: "Width", Value: int32(sz.Width)},
-		{Key: "Height", Value: int32(sz.Height)},
+		{Key: "Width", Value: int64(sz.Width)},
+		{Key: "Height", Value: int64(sz.Height)},
 	}
 }
 
@@ -611,7 +611,7 @@ func serializeTextTemplate(text *model.Text, params []string) bson.D {
 	}
 	if len(text.Translations) > 0 {
 		var transArray bson.A
-		transArray = append(transArray, int32(3)) // items marker (3 = has items)
+		transArray = append(transArray, int64(3)) // items marker (3 = has items)
 		for lang, value := range text.Translations {
 			transArray = append(transArray, bson.D{
 				{Key: "$ID", Value: idToBsonBinary(generateUUID())},
